requirements.txt
google-cloud-aiplatform>=1.87.0
google-cloud-bigquery>=3.25.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
httpx>=0.27.0
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.127.0
pyyaml>=6.0.1
python-dotenv>=1.0.1
pydantic>=2.7.0
